Absorb DeleteAndInsertTest into DeleteQueryTest and InsertQueryTest
